requirements.txt
streamlit>=1.28.0
google-generativeai>=0.5.0
feedparser>=6.0.0
biopython>=1.81
requests>=2.31.0
pandas>=2.0.0
notion-client>=2.0.0
